move navigation in view
