+++ Snabble/UI/ShopFinder/Model/ShopsViewModel.swift
@@ -9,7 +9,6 @@ import Foundation
 import Combine
 import CoreLocation
 import MapKit
-import Contacts
 
 /// ShopFinderViewModel for objects implermenting the ShopInfoProvider protocol
 public final class ShopsViewModel: NSObject, ObservableObject {
@@ -95,19 +94,6 @@ extension ShopsViewModel {
 
     static func region(for shop: ShopProviding) -> MKCoordinateRegion {
         MKCoordinateRegion(center: shop.location.coordinate, latitudinalMeters: 2000, longitudinalMeters: 2000)
-    }
-}
-
-extension ShopsViewModel {
-    static func navigate(to shop: ShopProviding) {
-        let endingItem = MKMapItem(placemark: MKPlacemark(coordinate: CLLocationCoordinate2DMake(shop.latitude, shop.longitude),
-                                                          addressDictionary: [
-                                                            CNPostalAddressCityKey: shop.city,
-                                                            CNPostalAddressStreetKey: shop.street,
-                                                            CNPostalAddressPostalCodeKey: shop.postalCode,
-                                                            CNPostalAddressISOCountryCodeKey: shop.country
-                                                          ]))
-        endingItem.openInMaps(launchOptions: [MKLaunchOptionsDirectionsModeKey: MKLaunchOptionsDirectionsModeDriving])
     }
 }
